refactor(#652): move explorer page to app router
