+++ frontend/src/app/(scoreboard)/games/explorer/page.tsx
@@ -1,0 +1,10 @@
+import { Suspense } from 'react';
+import ExplorerPage from './ExplorerPage';
+
+export default function Page() {
+    return (
+        <Suspense>
+            <ExplorerPage />
+        </Suspense>
+    );
+}
