+++ frontend/src/games/list/SearchFilters.tsx
@@ -1,5 +1,12 @@
+import { EventType, trackEvent } from '@/analytics/events';
+import { useApi } from '@/api/Api';
+import { useFreeTier, useRequiredAuth } from '@/auth/Auth';
 import { MastersCohort } from '@/database/game';
+import { RequirementCategory } from '@/database/requirement';
+import { dojoCohorts } from '@/database/user';
 import { SearchFunc } from '@/hooks/usePagination';
+import CohortIcon from '@/scoreboard/CohortIcon';
+import Icon from '@/style/Icon';
 import { Folder } from '@mui/icons-material';
 import ArrowForwardIosSharpIcon from '@mui/icons-material/ArrowForwardIosSharp';
 import { LoadingButton } from '@mui/lab';
@@ -23,18 +30,7 @@ import { styled } from '@mui/material/styles';
 import { DatePicker } from '@mui/x-date-pickers/DatePicker';
 import { DateTime } from 'luxon';
 import React, { useCallback, useEffect, useState } from 'react';
-import {
-    Link as RouterLink,
-    URLSearchParamsInit,
-    useSearchParams,
-} from 'react-router-dom';
-import { EventType, trackEvent } from '../../analytics/events';
-import { useApi } from '../../api/Api';
-import { useFreeTier, useRequiredAuth } from '../../auth/Auth';
-import { RequirementCategory } from '../../database/requirement';
-import { dojoCohorts } from '../../database/user';
-import CohortIcon from '../../scoreboard/CohortIcon';
-import Icon from '../../style/Icon';
+import { URLSearchParamsInit, useSearchParams } from 'react-router-dom';
 
 const Accordion = styled((props: AccordionProps) => (
     <MuiAccordion disableGutters elevation={0} square {...props} />
@@ -479,8 +475,7 @@ const SearchByPosition: React.FC<SearchByPositionProps> = ({
                 </Typography>
             ) : (
                 <LoadingButton
-                    component={RouterLink}
-                    to={`/games/explorer?fen=${fen}`}
+                    href={`/games/explorer?fen=${fen}`}
                     disabled={isLoading}
                     variant='outlined'
                     startIcon={<Icon name='explore' color='primary' />}
